[opensubdiv/hbr] Addressed compiler warnings

+++ opensubdiv/hbr/catmark.h
@@ -185,7 +185,7 @@ HbrCatmarkSubdivision<T>::transferFVarToChild(HbrMesh<T>* mesh, HbrFace<T>* face
         if (fvarinterp == HbrMesh<T>::k_InterpolateBoundaryNone ||
             (fvarinterp == HbrMesh<T>::k_InterpolateBoundaryAlwaysSharp &&
              fvarmask >= 1) ||
-            v->GetSharpness() > HbrVertex<T>::k_Smooth ||
+            v->GetSharpness() > (float) HbrVertex<T>::k_Smooth ||
             infcorner) {
             fv0.SetWithWeight(face->GetFVarData(index), fvarindex, fvarwidth, 1.0f);
         }
@@ -458,7 +458,7 @@ HbrCatmarkSubdivision<T>::Refine(HbrMesh<T>* mesh, HbrFace<T>* face) {
             // Hand down edge sharpnesses
             childedge = vertex->Subdivide()->GetEdge(edge->Subdivide());
             assert(childedge);
-            if ((sharpness = edge->GetSharpness()) > HbrHalfedge<T>::k_Smooth) {
+            if ((sharpness = edge->GetSharpness()) > (float) HbrHalfedge<T>::k_Smooth) {
                 HbrSubdivision<T>::SubdivideCreaseWeight(
                     edge, edge->GetOrgVertex(), childedge);
             }
@@ -466,7 +466,7 @@ HbrCatmarkSubdivision<T>::Refine(HbrMesh<T>* mesh, HbrFace<T>* face) {
 
             childedge = prevedge->Subdivide()->GetEdge(vertex->Subdivide());
             assert(childedge);
-            if ((sharpness = prevedge->GetSharpness()) > HbrHalfedge<T>::k_Smooth) {
+            if ((sharpness = prevedge->GetSharpness()) > (float) HbrHalfedge<T>::k_Smooth) {
                 HbrSubdivision<T>::SubdivideCreaseWeight(
                     prevedge, prevedge->GetDestVertex(), childedge);
             }
@@ -539,7 +539,7 @@ HbrCatmarkSubdivision<T>::RefineFaceAtVertex(HbrMesh<T>* mesh, HbrFace<T>* face,
                 // Hand down edge sharpness
                 childedge = vertex->Subdivide()->GetEdge(edge->Subdivide());
                 assert(childedge);
-                if ((sharpness = edge->GetSharpness()) > HbrHalfedge<T>::k_Smooth) {
+                if ((sharpness = edge->GetSharpness()) > (float) HbrHalfedge<T>::k_Smooth) {
                     HbrSubdivision<T>::SubdivideCreaseWeight(
                         edge, edge->GetOrgVertex(), childedge);
                 }
@@ -547,7 +547,7 @@ HbrCatmarkSubdivision<T>::RefineFaceAtVertex(HbrMesh<T>* mesh, HbrFace<T>* face,
 
                 childedge = prevedge->Subdivide()->GetEdge(vertex->Subdivide());
                 assert(childedge);
-                if ((sharpness = prevedge->GetSharpness()) > HbrHalfedge<T>::k_Smooth) {
+                if ((sharpness = prevedge->GetSharpness()) > (float) HbrHalfedge<T>::k_Smooth) {
                     HbrSubdivision<T>::SubdivideCreaseWeight(
                         prevedge, prevedge->GetDestVertex(), childedge);
                 }
@@ -771,7 +771,7 @@ HbrCatmarkSubdivision<T>::HasLimit(HbrMesh<T>* mesh, HbrHalfedge<T>* edge) {
     //  A smooth edge has a limit if both endpoints have limits and
     //  the edge isn't on the boundary.
 
-    if (edge->GetSharpness() >= HbrHalfedge<T>::k_InfinitelySharp) return true;
+    if (edge->GetSharpness() >= (float) HbrHalfedge<T>::k_InfinitelySharp) return true;
 
     if (!HasLimit(mesh, edge->GetOrgVertex()) || !HasLimit(mesh, edge->GetDestVertex())) return false;
 
@@ -795,7 +795,7 @@ HbrCatmarkSubdivision<T>::HasLimit(HbrMesh<T>* /* mesh */, HbrVertex<T>* vertex)
                 HbrHalfedge<T>* start = vertex->GetIncidentEdge(), *edge, *next;
                 edge = start;
                 while (edge) {
-                    if (edge->IsBoundary() && edge->GetSharpness() < HbrHalfedge<T>::k_InfinitelySharp) {
+                    if (edge->IsBoundary() && edge->GetSharpness() < (float) HbrHalfedge<T>::k_InfinitelySharp) {
                         return false;
                     }
                     next = vertex->GetNextEdge(edge);
@@ -803,7 +803,7 @@ HbrCatmarkSubdivision<T>::HasLimit(HbrMesh<T>* /* mesh */, HbrVertex<T>* vertex)
                         break;
                     } else if (!next) {
                         edge = edge->GetPrev();
-                        if (edge->IsBoundary() && edge->GetSharpness() < HbrHalfedge<T>::k_InfinitelySharp) {
+                        if (edge->IsBoundary() && edge->GetSharpness() < (float) HbrHalfedge<T>::k_InfinitelySharp) {
                             return false;
                         }
                         break;
@@ -1092,9 +1092,9 @@ HbrCatmarkSubdivision<T>::Subdivide(HbrMesh<T>* mesh, HbrVertex<T>* vertex) {
     // Inherit extraordinary flag and sharpness
     if (vertex->IsExtraordinary()) v->SetExtraordinary();
     float sharp = vertex->GetSharpness();
-    if (sharp >= HbrVertex<T>::k_InfinitelySharp) {
+    if (sharp >= (float) HbrVertex<T>::k_InfinitelySharp) {
         v->SetSharpness(HbrVertex<T>::k_InfinitelySharp);
-    } else if (sharp > HbrVertex<T>::k_Smooth) {
+    } else if (sharp > (float) HbrVertex<T>::k_Smooth) {
         sharp -= 1.0f;
         if (sharp < (float) HbrVertex<T>::k_Smooth) {
             sharp = (float) HbrVertex<T>::k_Smooth;
